Refactor: Move private functions to public
